+++ buildconfig/stubs/pygame/event.pyi
@@ -1,10 +1,19 @@
-from typing import Any, Union, Optional
+from typing import Any, Union, Optional, Dict
 
 from pygame.typing import SequenceLike, EventLike
 
 
 class Event(EventLike):
+    def __init__(
+        self, type: int, dict: Optional[Dict[str, Any]] = None, **kwargs: Any
+    ) -> None: ...
     def __new__(cls, *args: Any, **kwargs: Any) -> "Event": ...
+    def __getattribute__(self, name: str) -> Any: ...
+    def __setattr__(self, name: str, value: Any) -> None: ...
+    def __delattr__(self, name: str) -> None: ...
+    def __int__(self) -> int: ...
+    def __bool__(self) -> bool: ...
+    def __eq__(self, other: Any) -> bool: ...
 
 _EventTypes = Union[int, SequenceLike[int]]
