+++ buildconfig/stubs/pygame/typing.pyi
@@ -84,15 +84,9 @@ RectLike = Union[SequenceLike[float], SequenceLike[Coordinate], _HasRectAttribut
 
 class EventLike(Protocol):
     def __init__(
-        self, type: int, dict: Optional[Dict[str, Any]] = None, **kwargs: Any
+        self, type: int, /, **kwargs: Any
     ) -> None: ...
     def __new__(cls, *args: Any, **kwargs: Any) -> "EventLike": ...
-    def __getattribute__(self, name: str) -> Any: ...
-    def __setattr__(self, name: str, value: Any) -> None: ...
-    def __delattr__(self, name: str) -> None: ...
-    def __int__(self) -> int: ...
-    def __bool__(self) -> bool: ...
-    def __eq__(self, other: Any) -> bool: ...
 
     @property
     def type(self) -> int: ...
