+++ src_c/doc/typing_doc.h
@@ -8,3 +8,4 @@
 #define DOC_TYPING_RGBATUPLE ""
 #define DOC_TYPING_COLORLIKE ""
 #define DOC_TYPING_RECTLIKE ""
+#define DOC_TYPING_EVENTLIKE ""
